requirements.txt
asgiref==3.8.1
Django==5.1.7
django-taggit==6.1.0
Markdown==3.7
psycopg==3.1.18
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
